Move duckdb Parquet reader to separate folder (#1767)
Summary:
This PR is to prepare for the upcoming native Parquet reader. It moves the duckdb Parquet reader into dwio/parquet/duckdb_reader. Users will be able to choose duckdb Parquet reader or the new native Parquet reader in the future.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1767

Reviewed By: Yuhta

Differential Revision: D37163205

Pulled By: zzhao0

fbshipit-source-id: af6fcb2fadff496336c04610bfbd8f42d4fdcaa9
