+++ velox/dwio/parquet/CMakeLists.txt
@@ -12,8 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_subdirectory(reader)
+add_subdirectory(duckdb_reader)
 
 if(${VELOX_BUILD_TESTING})
   add_subdirectory(tests)
 endif()
+
+add_library(velox_dwio_parquet_reader RegisterParquetReader.cpp)
+target_link_libraries(velox_dwio_parquet_reader
+                      velox_dwio_duckdb_parquet_reader gtest xsimd)
